+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Release date: ?
  - adding support of curly braces syntax for route parameters in Http\Request\Route 
  - added Http\Request\RouteParameter and updated Http\Request\Route to use it
  - added Http\Request\RouteServerRequest
+ - added Common\Trait\MicroTime
+ - refactored Common\Chrono
 
 VERSION 4.0-BETA2
 -----------